move ProductInfo into separate file
